Move dependency graph code back into core module.
